+++ v3/src/js/views/timetable/ModulesSelect.jsx
